Fix pkg2appimage tool download URL
I spotted two issues. 

First one is that the download URL for the pkg2appimage tool that is used to create an appimage from the debian package was not valid anymore as the pkg2appimge project hast been moved to a different GitHub organization (now AppImageCommunity).

Second one is that the dynamic way to get the latest version of the tool (which is also described on their README.md) does not work anymore as GitHub seems to have change the way assets for a release are being displayed. Until now you could simply `wget` the releases page and the assets to each release would be in there. Now when I open the releases page there is a loading spinner where the assets for a particular release should be. They are being lazy-loaded.

So for now I'm hard-coding the pkg2appimage URL until I find a solution to do it dynamically again. Maybe via the GitHub API or CLI.